pyproject.toml
[tool.poetry]
name = "yap_auth"
version = "0.1.0"
description = ""
authors = ["Maxim B. <9652485@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "2.1.2"
Flask-SQLAlchemy = "2.5.1"
flask-swagger-ui = "3.36.0"
aioredis = "2.0.1"
psycopg2 = "2.9.3"
pydantic = "1.9.0"
python-dotenv = "0.20.0"
flask-restx = "0.5.1"
flask-marshmallow = "0.14.0"
marshmallow-sqlalchemy = "0.28.0"

[tool.poetry.dev-dependencies]
black = {version = "^22.3.0", allow-prereleases = true}
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py39']
include = '\.pyi?$'
exclude =  '''
/(
    \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.pytest_cache
    | \.tox
    | \.vscode
    | __pycache__
    | _build
    | buck-out
    | build
    | dist
    | venv
)/
'''

[tool.isort]
profile = "black"